Restructure blocks.py files
